Split e2e tests into modules #270
