+++ lock-keeper-tests/src/end_to_end/operations/authenticate.rs
@@ -1,0 +1,16 @@
+use lock_keeper_client::{client::Password, LockKeeperClient};
+
+use crate::end_to_end::Test;
+
+impl Test {
+    pub async fn authenticate(&self, password: &Option<Password>) -> anyhow::Result<()> {
+        let password = match password {
+            Some(pwd) => pwd,
+            None => &self.password,
+        };
+        let _ = LockKeeperClient::authenticated_client(&self.account_name, password, &self.config)
+            .await?;
+
+        Ok(())
+    }
+}
